So much... Easiest is just to dump here the ChangeLog entries (which are imprecise as well..)
1998-07-13  Matti Aarnio  <mea@mea.tmt.tele.fi>
* utils/policy-builder.sh.in:
    Mark Visser <mark@cal026031.student.utwente.nl> reported
    about the clumsiness of the spam-list retrieve and merge
    code.  His patch went in almoast as is.
* aclocal.m4, transports/mailbox/mailbox.c:
    Debian hits again; I don't know which version this is about,
    but some recent ones have <maillock.h> header, and library
    libfilelock.{a,so} with similar code to that of the Solaris
    libmail.a's maillock() stuff.
* proto/cf/rrouter.cf:
    Move the "new localpart processing" a bit downwards so that
    "pipe.*" and "file.*" tricks used with the shceduler configs
    still work.

1998-07-08  Matti Aarnio  <mea@mea.tmt.tele.fi>
* */*.h */*.c */*/*.c: (all over the place)
    Did use "egcs -Wall -W -W***" to get lots of warnings.
    Tried to fix most of them.
* transports/mailbox/mailbox.c, transports/sm/sm.c:
    Internally the  append_header()  routine has a buffer of 2kB
    size.  Under no circumstances allow that to be exceeded, thus
    use patterns like:  "%s.999s"  while formatting something.

+++ lib/emalloc.c
@@ -39,7 +39,7 @@ erealloc(buf, len)
 	while ((r = realloc(buf, len)) == NULL) {
 		fprintf(stderr,
 			"%s[%d] realloc(%u): virtual memory exceeded, sleeping\n",
-			progname, getpid(), len);
+			progname, getpid(), (u_int)len);
 		sleep(emsleeptime);
 	}
 	return r;
